- Rename project to EasyVCR - Add docstrings - Add AutoHandler (tries to replay, fallback to record when missing) - Option to use strict matching (match body as well as method and URI) - Matching converts URI and body to base64 string to improve matching performance
