+++ EasyPost.EasyVCR/VCRException.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace EasyPost.EasyVCR
+{
+    public class VCRException : Exception
+    {
+        internal VCRException(string message) : base(message)
+        {
+        }
+    }
+}
