--- EasyPost.Scotch.Tests/EasyPost.Scotch.Tests.csproj
@@ -1,26 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<Project Sdk="Microsoft.NET.Sdk.Web">
-    <PropertyGroup>
-        <TargetFrameworks>net461;netstandard2.0;netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
-        <LangVersion>9</LangVersion>
-        <ApplicationIcon />
-        <OutputType>Library</OutputType>
-        <StartupObject />
-        <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
-        <RootNamespace>EasyPost.Scotch.Tests</RootNamespace>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-    <ItemGroup>
-        <ProjectReference Include="..\EasyPost.Scotch\EasyPost.Scotch.csproj" />
-    </ItemGroup>
-    <ItemGroup>
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
-        <PackageReference Include="MSTest.TestAdapter" Version="2.2.8" />
-        <PackageReference Include="MSTest.TestFramework" Version="2.2.8" />
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-    </ItemGroup>
-    <ItemGroup>
-      <Folder Include="cassettes" />
-    </ItemGroup>
-</Project>
