--- EasyPost.Scotch/InternalUtilities/JSON/Orders/OrderOption.cs
@@ -1,11 +1,0 @@
-using System;
-using System.Collections.Generic;
-using Newtonsoft.Json.Serialization;
-
-namespace EasyPost.Scotch.InternalUtilities.JSON.Orders
-{
-    public interface IOrderOption
-    {
-        internal Func<IList<JsonProperty>, IEnumerable<JsonProperty>>? OrderFunction { get; }
-    }
-}
