--- EasyPost.Scotch/InternalUtilities/Utils.cs
@@ -1,12 +1,0 @@
-using System.IO;
-
-namespace EasyPost.Scotch.InternalUtilities
-{
-    internal static class Utils
-    {
-        internal static string GetFilePath(string folderPath, string fileName)
-        {
-            return Path.Combine(folderPath, fileName);
-        }
-    }
-}
