--- EasyPost.Scotch/RequestElements/HttpInteraction.cs
@@ -1,15 +1,0 @@
-using System;
-using Newtonsoft.Json;
-
-namespace EasyPost.Scotch.RequestElements
-{
-    internal class HttpInteraction
-    {
-        [JsonProperty("RecordedAt")]
-        internal DateTimeOffset RecordedAt { get; set; }
-        [JsonProperty("Request")]
-        internal Request Request { get; set; }
-        [JsonProperty("Response")]
-        internal Response Response { get; set; }
-    }
-}
