--- EasyPost.Scotch/ScotchMode.cs
@@ -1,9 +1,0 @@
-namespace EasyPost.Scotch
-{
-    public enum ScotchMode
-    {
-        None = 0,
-        Recording = 1,
-        Replaying = 2
-    }
-}
