--- EasyPost.Scotch/VCRException.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace EasyPost.Scotch
-{
-    public class VCRException : Exception
-    {
-        internal VCRException(string message) : base(message)
-        {
-        }
-    }
-}
