+++ EasyVCR.sln
@@ -1,0 +1,36 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25123.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EasyPost.EasyVCR", "EasyPost.EasyVCR\EasyPost.EasyVCR.csproj", "{548240D7-3DDC-4256-A3EB-86FB249E3375}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EasyPost.EasyVCR.Tests", "EasyPost.EasyVCR.Tests\EasyPost.EasyVCR.Tests.csproj", "{F44D9A65-279B-47B9-8CF0-11CB3ACC4370}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E0A16E71-5E47-449D-97C1-D9EF205E2668}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E0A16E71-5E47-449D-97C1-D9EF205E2668}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E0A16E71-5E47-449D-97C1-D9EF205E2668}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E0A16E71-5E47-449D-97C1-D9EF205E2668}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D5B95235-8B0E-4A66-96E7-FA79116AE596}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D5B95235-8B0E-4A66-96E7-FA79116AE596}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D5B95235-8B0E-4A66-96E7-FA79116AE596}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D5B95235-8B0E-4A66-96E7-FA79116AE596}.Release|Any CPU.Build.0 = Release|Any CPU
+		{548240D7-3DDC-4256-A3EB-86FB249E3375}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{548240D7-3DDC-4256-A3EB-86FB249E3375}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{548240D7-3DDC-4256-A3EB-86FB249E3375}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{548240D7-3DDC-4256-A3EB-86FB249E3375}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F44D9A65-279B-47B9-8CF0-11CB3ACC4370}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F44D9A65-279B-47B9-8CF0-11CB3ACC4370}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F44D9A65-279B-47B9-8CF0-11CB3ACC4370}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F44D9A65-279B-47B9-8CF0-11CB3ACC4370}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
